Rename peer_directory -> directory
