+++ elements/lisk-p2p/src/p2p.ts
@@ -30,7 +30,7 @@ import {
 	REMOTE_RPC_GET_PEERS_LIST,
 } from './peer';
 
-import { PeerBook } from './peer_directory';
+import { PeerBook } from './directory';
 
 import {
 	FORBIDDEN_CONNECTION,
