+++ elements/lisk-p2p/test/unit/peer_directory/new_peers.ts
@@ -13,7 +13,7 @@
  *
  */
 import { expect } from 'chai';
-import { NewPeers } from '../../../src/peer_directory/new_peers';
+import { NewPeers } from '../../../src/directory/new_peers';
 import { initializePeerInfoList } from '../../utils/peers';
 import { P2PPeerInfo } from '../../../src/p2p_types';
