+++ elements/lisk-p2p/test/unit/peer_directory/peer_book.ts
@@ -16,10 +16,7 @@ import { expect } from 'chai';
 import { initializePeerInfoList } from '../../utils/peers';
 // import { constructPeerIdFromPeerInfo } from '../../../src/utils';
 // import { P2PPeerInfo } from '../../../src/p2p_types';
-import {
-	PeerBook,
-	PeerBookConfig,
-} from '../../../src/peer_directory/peer_book';
+import { PeerBook, PeerBookConfig } from '../../../src/directory/peer_book';
 
 describe('peerBook', () => {
 	const peerBookConfig: PeerBookConfig = {
